Update the regexp for Google search results
Some searches have a different class of <h3>, which contains the title
and URL of a result. This different class seems to be used when a
result gets 'promoted' to appear prominently before the other results.
We also want this results, since it doesn't appear afterwards in the
normal results.