MeidoBot: Fix bug in Main, also assign thread names
On the thread names front, it seems like it works kinda strange. If you
assign a thread name before starting a thread it doesn't seem to take,
but if you do it after you start the thread it does. (see LogWriter)
Thread.CurrentThread.Name works fine as well, but that makes sense -
cause to set that you have to be in the relevant thread and it has to be
running (ie started).