IrcCalc: Fix 2 small but important bugs
In Tokenizer I neglected to record the start of a number when it started
with a dot/decimal point. This made a mess for all subsequent number
characters.

In ShuntingYard I completely missed the fact that because it's a stack
based solution the function arguments will have to be read 'backwards'.
Last in, First out.