Lay the groundwork for event handling rework
The idea is that plugins will have to declare their handlers in a
IEnumerable<IIrcHandler> field. The IrcHandler<T> class is the defacto
implementation of that interface. We do it this way because we have to
abstract over the various IRC event types, the type system is not
expressive enough to do what we actually want:
IEnumerable<IrcHandler<T>> where T is unbound.