MeidoCommon: Add new help system
HelpRequest parses the help query. HelpResult takes care of formatting
the first line and doling out the remaining help in an acceptable way.
The main Help class decides what to look up (dependent on HelpRequest)
and formats the footer. This formatting is quite lengthy and might be a
good candidate for seperating into another class.