Chainey: Use SqliteConnection directly
Instead of going through all the ThreadLocal bussiness. Yes, this makes
it not thread-safe (probably), cause that's why I did it like this in
the first place. But since I'm changing upstream to be single threaded
we'll just have to accept that this class will not be thread-safe.

Also change to nameof where appropriate.