NyaaSpam: Fix rescheduling bug
When the logic was changed to reschedule after every read I missed that
after an error getting and/or opening the feed the rest of the function
wouldn't get run, including the rescheduling.

Having added another level of indentation I would like to split this
function, but most parts need access to various bookkeeping variables,
so it's all kinda entangled.