NyaaSpam: Use ArgEnumerator
Most parsing has gotten easier with the new methods, but NyaaSpam has
always been beast. In fact, the ArgEnumerator idea started with me
thinking on how to best tackle the more complex arguments and branching
thereof that NyaaSpam exemplified (as far as my bot plugins go). So I'm
happy to see that it indeed leads to simpler, more concise code. Not
having to manually kludge with arrays and indices is a relief.